requirements.txt
streamlit
PyMuPDF
requests
Pillow
python-dotenv
tiktoken
nltk
openai
numpy
python-docx
